+++ package/telegram_client/lib/tdlib/tdlib_ffi/tdlib_io.dart
@@ -495,7 +495,7 @@ class LibTdJson {
       }
     }
     if (extra_id.isEmpty) {
-      extra_id= generateUuid(15);
+      extra_id = generateUuid(15);
       parameters["@extra"] = extra_id;
     }